refactor(core,toolkit): extract connector loading code to toolkit
