+++ packages/core/src/index.ts
@@ -9,7 +9,7 @@ import SystemContext from './tenants/SystemContext.js';
 dotenv.config({ path: await findUp('.env', {}) });
 
 // Import after env has been configured
-const { loadConnectorFactories } = await import('./utils/connectors/factories.js');
+const { loadConnectorFactories } = await import('./utils/connectors/index.js');
 const { EnvSet } = await import('./env-set/index.js');
 const { default: initI18n } = await import('./i18n/init.js');
 const { tenantPool, checkRowLevelSecurity } = await import('./tenants/index.js');
