+++ packages/phrases-ui/src/index.ts
@@ -1,6 +1,5 @@
-import { fallback } from '@logto/core-kit';
 import type { LanguageTag } from '@logto/language-kit';
-import { languages } from '@logto/language-kit';
+import { languages, fallback } from '@logto/language-kit';
 import type { NormalizeKeyPaths } from '@silverhand/essentials';
 import { z } from 'zod';
