+++ packages/phrases/package.json
@@ -33,7 +33,6 @@
     "url": "https://github.com/logto-io/logto/issues"
   },
   "dependencies": {
-    "@logto/core-kit": "workspace:*",
     "@logto/language-kit": "workspace:*",
     "@silverhand/essentials": "2.4.0",
     "zod": "^3.20.2"
