+++ packages/toolkit/core-kit/src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './color.js';
 export * from './id.js';
-export * from './zod.js';
 export * from './url.js';
