+++ packages/toolkit/language-kit/src/utility.test.ts
@@ -1,4 +1,6 @@
-import { isLanguageTag, languageTagGuard } from './utility.js';
+import { number, ZodError } from 'zod';
+
+import { fallback, isLanguageTag, languageTagGuard } from './utility.js';
 
 describe('isLanguageTag', () => {
   it('should pass when input is a valid language key', () => {
@@ -23,5 +25,15 @@ describe('languageTagGuard', () => {
     for (const invalidLanguageKey of [undefined, '', 'xx-XX']) {
       expect(languageTagGuard.safeParse(invalidLanguageKey).success).toBeFalsy();
     }
+  });
+});
+
+describe('fallback', () => {
+  it('should fallback to default value', () => {
+    const schema = number();
+    const tolerant = schema.or(fallback(-1));
+
+    expect(() => schema.parse('foo')).toThrow(ZodError);
+    expect(tolerant.parse('foo')).toBe(-1);
   });
 });
