+++ packages/toolkit/language-kit/src/utility.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z, any } from 'zod';
 
 import { languages } from './const.js';
 import type { LanguageTag } from './type.js';
@@ -9,3 +9,17 @@ export const isLanguageTag = (value: unknown): value is LanguageTag =>
 export const languageTagGuard: z.ZodType<LanguageTag> = z
   .any()
   .refine((value: unknown) => isLanguageTag(value));
+
+/**
+ * https://github.com/colinhacks/zod/issues/316#issuecomment-850906479
+ * Create a schema matches anything and returns a value. Use it with `or`:
+ *
+ * const schema = zod.number();
+ * const tolerant = schema.or(fallback(-1));
+ *
+ * schema.parse('foo')      // => ZodError
+ * tolerant.parse('foo')    // -1
+ */
+export function fallback<T>(value: T) {
+  return any().transform(() => value);
+}
